# src/OnlineBookShop.React/src/shared/api/generated/models/BookForUpdateDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OnlineBookShop.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BookForUpdateDto
 */
export interface BookForUpdateDto {
    /**
     * 
     * @type {string}
     * @memberof BookForUpdateDto
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof BookForUpdateDto
     */
    description?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof BookForUpdateDto
     */
    publishedOn: Date;
    /**
     * 
     * @type {number}
     * @memberof BookForUpdateDto
     */
    publisherId: number;
    /**
     * 
     * @type {number}
     * @memberof BookForUpdateDto
     */
    price: number;
}

export function BookForUpdateDtoFromJSON(json: any): BookForUpdateDto {
    return BookForUpdateDtoFromJSONTyped(json, false);
}

export function BookForUpdateDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BookForUpdateDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'title': json['title'],
        'description': !exists(json, 'description') ? undefined : json['description'],
        'publishedOn': (new Date(json['publishedOn'])),
        'publisherId': json['publisherId'],
        'price': json['price'],
    };
}

export function BookForUpdateDtoToJSON(value?: BookForUpdateDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'title': value.title,
        'description': value.description,
        'publishedOn': (value.publishedOn.toISOString()),
        'publisherId': value.publisherId,
        'price': value.price,
    };
}