# src/OnlineBookShop.React/src/shared/api/generated/models/BookGridRowDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OnlineBookShop.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BookGridRowDto
 */
export interface BookGridRowDto {
    /**
     * 
     * @type {number}
     * @memberof BookGridRowDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof BookGridRowDto
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BookGridRowDto
     */
    description?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof BookGridRowDto
     */
    publishedOn?: Date;
    /**
     * 
     * @type {string}
     * @memberof BookGridRowDto
     */
    publisher?: string | null;
    /**
     * 
     * @type {number}
     * @memberof BookGridRowDto
     */
    price?: number;
}

export function BookGridRowDtoFromJSON(json: any): BookGridRowDto {
    return BookGridRowDtoFromJSONTyped(json, false);
}

export function BookGridRowDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BookGridRowDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'title': !exists(json, 'title') ? undefined : json['title'],
        'description': !exists(json, 'description') ? undefined : json['description'],
        'publishedOn': !exists(json, 'publishedOn') ? undefined : (new Date(json['publishedOn'])),
        'publisher': !exists(json, 'publisher') ? undefined : json['publisher'],
        'price': !exists(json, 'price') ? undefined : json['price'],
    };
}

export function BookGridRowDtoToJSON(value?: BookGridRowDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'title': value.title,
        'description': value.description,
        'publishedOn': value.publishedOn === undefined ? undefined : (value.publishedOn.toISOString()),
        'publisher': value.publisher,
        'price': value.price,
    };
}