# src/OnlineBookShop.React/src/shared/api/generated/models/PagedRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OnlineBookShop.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    RequestFilters,
    RequestFiltersFromJSON,
    RequestFiltersFromJSONTyped,
    RequestFiltersToJSON,
} from './';

/**
 * 
 * @export
 * @interface PagedRequest
 */
export interface PagedRequest {
    /**
     * 
     * @type {number}
     * @memberof PagedRequest
     */
    pageIndex?: number;
    /**
     * 
     * @type {number}
     * @memberof PagedRequest
     */
    pageSize?: number;
    /**
     * 
     * @type {string}
     * @memberof PagedRequest
     */
    columnNameForSorting?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PagedRequest
     */
    sortDirection?: string | null;
    /**
     * 
     * @type {RequestFilters}
     * @memberof PagedRequest
     */
    requestFilters?: RequestFilters;
}

export function PagedRequestFromJSON(json: any): PagedRequest {
    return PagedRequestFromJSONTyped(json, false);
}

export function PagedRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PagedRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'pageIndex': !exists(json, 'pageIndex') ? undefined : json['pageIndex'],
        'pageSize': !exists(json, 'pageSize') ? undefined : json['pageSize'],
        'columnNameForSorting': !exists(json, 'columnNameForSorting') ? undefined : json['columnNameForSorting'],
        'sortDirection': !exists(json, 'sortDirection') ? undefined : json['sortDirection'],
        'requestFilters': !exists(json, 'requestFilters') ? undefined : RequestFiltersFromJSON(json['requestFilters']),
    };
}

export function PagedRequestToJSON(value?: PagedRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'pageIndex': value.pageIndex,
        'pageSize': value.pageSize,
        'columnNameForSorting': value.columnNameForSorting,
        'sortDirection': value.sortDirection,
        'requestFilters': RequestFiltersToJSON(value.requestFilters),
    };
}