# src/OnlineBookShop.React/src/shared/api/generated/models/PublisherDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OnlineBookShop.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PublisherDto
 */
export interface PublisherDto {
    /**
     * 
     * @type {number}
     * @memberof PublisherDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof PublisherDto
     */
    name?: string | null;
}

export function PublisherDtoFromJSON(json: any): PublisherDto {
    return PublisherDtoFromJSONTyped(json, false);
}

export function PublisherDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PublisherDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'name': !exists(json, 'name') ? undefined : json['name'],
    };
}

export function PublisherDtoToJSON(value?: PublisherDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
    };
}